% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hrspline.R
\name{plotHR}
\alias{plotHR}
\title{Plot the result of an object of class HR}
\usage{
plotHR(
  x,
  xlab = "",
  main = "",
  log = FALSE,
  ylab = if (log) "log(HR)" else "HR",
  line1 = TRUE,
  color = "dodgerblue"
)
}
\arguments{
\item{x}{data.frame calculated using HRcubSpline}

\item{xlab}{xlab name}

\item{main}{plot title}

\item{log}{if TRUE, plot HR in log scale}

\item{ylab}{ylab name. Default HR if log=FALSE otherwise HR(log scale)}

\item{line1}{if TRUE, plot horizontal line on HR = 1 or log(HR) = 0}

\item{color}{HR line color. Default dodgerblue}
}
\value{
simple splined plot of your var2 values by HR
}
\description{
Create a spline var2 by HR of var1
}
\examples{
library(rms)
library(survival)
data(cancer)
myformula <- Surv(time, status) ~ ph.karno + ph.ecog + rcs(age, 3)*sex
model <- cph(myformula , data = lung )
myHR <- rcsHR( var2values = 40:80
                     , model = model , data = lung , var1 ="sex", var2="age"
                     , ci=TRUE , conf = 0.95 , ci.method = "delta")
plotHR(myHR , ylab = "HR of male VS female" , xlab = "Age")
}
