% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcubspline.R
\name{rcsOR}
\alias{rcsOR}
\title{Restricted cubic spline interaction OR for more than 3 knots}
\usage{
rcsOR(
  var2values,
  model,
  data = NULL,
  var1,
  var2,
  ci = TRUE,
  conf = 0.95,
  ci.method = "delta",
  ci.boot.method = "perc",
  R = 100,
  parallel = "multicore",
  ...
)
}
\arguments{
\item{var2values}{numeric vector of var2 points to estimate}

\item{model}{model of class lrm, Glm or glm (family binomial, poisson or quasipoisson). If data is NULL, the function expects to find the data in model$x.}

\item{data}{data used in the model. If absent, we will attempt to recover the data from the model. Only used for bootstrap and glm class models}

\item{var1}{variable that increases by 1 unit from 0}

\item{var2}{variable to spline. var2values belong to var2}

\item{ci}{calculate 95\% CI?}

\item{conf}{confidence level. Default 0.95}

\item{ci.method}{confidence interval method. "delta" performs delta method. "bootstrap" performs bootstrapped CI (slower)}

\item{ci.boot.method}{one of the available bootstrap CI methods from \code{\link[boot]{boot.ci}}. Default percentile}

\item{R}{number of bootstrap samples if ci.method = "bootstrap". Default 100}

\item{parallel}{can take values "no", "multicore", "snow" if ci.method = "bootstrap". Default multicore}

\item{...}{other parameters for boot}
}
\value{
if ci = FALSE, a dataframe with initial values and OR (RR for Poisson models)
}
\description{
Generate OR values in a logistic model for a 1 unit increase in a variable at
specified points of another interacting variable splined with rcs(df >= 3).
For Poisson (or quasipoisson) models of class glm or Glm, rate ratios (RR) are returned instead.
}
\examples{
library(rms)
library(mlbench)
data(SynthDiabetes)
# Set age on a 5-year scale
SynthDiabetes$age <- SynthDiabetes$age/5
# Recode diabetes as 0/1
SynthDiabetes$diabetes <- ifelse(SynthDiabetes$diabetes=="pos" , 1 , 0)
myformula <- diabetes ~ mass + age * rcs( glucose , 4 )
model <- glm(myformula , data = SynthDiabetes , family = "binomial")
rcsOR( var2values = 20:80
       , model = model , data = SynthDiabetes , var1 ="age", var2="glucose"
       , ci=TRUE , conf = 0.95 , ci.method = "delta")
# Poisson model for a count outcome returns rate ratios (RR)
model_pois <- glm(pregnant ~ mass + age * rcs( glucose , 4 )
                  , data = SynthDiabetes , family = poisson())
rcsOR( var2values = 20:80
       , model = model_pois , data = SynthDiabetes , var1 ="age", var2="glucose"
       , ci=TRUE , conf = 0.95 , ci.method = "delta")
}
