% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umaru.R
\docType{data}
\name{umaru}
\alias{umaru}
\title{UMARU IMPACT Study data}
\format{
A data frame with 575 rows and 10 variables
\describe{
\item{X}{observation count}
\item{id}{identification code}
\item{age}{Age in years}
\item{beckscore}{Beck Depression Score at admission}
\item{ndrugtx}{Number of prior drug treatments}
\item{treat}{treat}
\item{site}{site}
\item{los}{los}
\item{time}{time to event}
\item{censor}{censor event}
\item{heroin}{heroin use at admission}
\item{cocaine}{cocaine use at admission}
\item{drugtx}{Prior Drug treatment, yes/no}
\item{ivdrug}{Prior IV drug treatment}
\item{nonwhite}{1 if non white, 0 otherwise}
}
}
\source{
\url{https://ftp.idu.ac.id/wp-content/uploads/ebook/ip/REGRESI\%20LOGISTIK/epdf.pub_applied-logistic-regression-wiley-series-in-probab.pdf}
}
\usage{
umaru
}
\description{
A subset of data from the University of Massachusets Aids Research Unit (UMARU) IMPACT study.
}
\keyword{datasets}
